requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
langchain_text_splitters
sentence_transformers
langchain_huggingface
langchain_core
faiss-cpu
langchain_groq
pypdf
streamlit